Change type of the attribute id of the book entity from Long to Integer

+++ src/main/java/com/apispring/apispring/application/dto/RequestUpdateBookDto.java
@@ -2,20 +2,20 @@ package com.apispring.apispring.application.dto;
 
 public class RequestUpdateBookDto extends BookDto {
 
-    private Long id;
+    private Integer id;
 
     public RequestUpdateBookDto(){ }
 
-    public RequestUpdateBookDto(Long id, String title, String author, String overview, Integer pages) {
+    public RequestUpdateBookDto(Integer id, String title, String author, String overview, Integer pages) {
         super(title, author, overview, pages);
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
